+++ src/Interest_Earning.cpp
@@ -8,7 +8,7 @@
 
 //p = Amount , r = rate of interest , t = time period 
 
-float Interest_Earning::computeInterest(float p , float r, int t)
+float Interest_Earning::computeInterest(float p , float r, float t)
 {
   std::cout << t << "\n";
   return p*pow((1+r/100),t);
